+++ mojo-descriptor-extractor/src/main/clojure/clojure/maven/extractor/clojure_mojo_extractor.clj
@@ -8,7 +8,7 @@
    [clojure.pprint :as pprint]
    classlojure)
   (:import
-   [clojure.maven.annotations Goal RequiresDependencyResolution]
+   [clojure.maven.annotations Goal Phase RequiresDependencyResolution]
    [org.apache.maven.plugin.descriptor
     MojoDescriptor
     Parameter
@@ -44,7 +44,7 @@
 
 (defn to-ns
   "Converts found mojo's into literal descriptor maps"
-  [base-path cl ^java.io.File file]
+  [base-path output-path cl ^java.io.File file]
   (when (and (.isFile file) (.endsWith (.getPath file) ".clj"))
     (let [path (.getPath file)
           path (.substring path (inc (count base-path)) (count path))
@@ -80,12 +80,14 @@
               mojo-descriptor#
               (fn [^java.lang.Class v#]
                 (let [^Goal goal# (.getAnnotation v# Goal)
+                      ^Phase phase# (.getAnnotation v# Phase)
                       ^RequiresDependencyResolution resolution#
                       (.getAnnotation v# RequiresDependencyResolution)]
                   (when-not goal#
                     (throw (IllegalArgumentException.
                             (str "No Goal annotation for " v#))))
-                  {:goal (.value goal#)
+                  {:goal (when goal# (.value goal#))
+                   :phase (when phase# (.value phase#))
                    :resolution (when resolution# (.value resolution#))
                    :implementation (.getName v#)
                    :fields (->>
@@ -100,9 +102,11 @@
                          (isa? v# org.apache.maven.plugin.Mojo))
                 (flush)
                 v#)]
+          (binding [*compile-path* ~output-path]
+            (compile '~try-ns))
           (require '~try-ns)
           (->>
-           (ns-map (find-ns '~try-ns))
+           (ns-map '~try-ns)
            (map second)
            (filter
             #(and (not (var? %))
@@ -112,10 +116,10 @@
            doall))))))
 
 (defn process-source-tree
-  [cl ^String source]
+  [output-path cl ^String source]
   (let [dir (java.io.File. source)]
     (->>
-     (mapcat #(to-ns (.getPath dir) cl %) (file-seq dir))
+     (mapcat #(to-ns (.getPath dir) output-path cl %) (file-seq dir))
      (filter identity))))
 
 (defn sources
@@ -156,7 +160,7 @@
     p))
 
 (defn mojo-descriptor
-  [plugin-descriptor {:keys [implementation goal resolution fields] :as m}]
+  [plugin-descriptor {:keys [implementation goal resolution fields phase]}]
   (when-not goal
     (throw
      (IllegalArgumentException.
@@ -166,6 +170,7 @@
                      (.setLanguage "clojure")
                      (.setImplementation implementation)
                      (.setGoal goal)
+                     (.setPhase (blank-as-nil phase))
                      (.setThreadSafe false))]
     (when resolution
       (.setDependencyResolutionRequired descriptor resolution))
@@ -178,13 +183,17 @@
 (defn plugin-classes
   [plugin-descriptor project]
   (let [sources (distinct (sources project))
+        output-path (java.io.File. (.. project getBuild getOutputDirectory))
         cl (apply
             classlojure/classlojure
             (map
              filename-to-url-string
              (concat sources (.getCompileClasspathElements project))))]
+    (when-not (.exists output-path)
+      (.mkdirs output-path))
     (java.util.ArrayList.
      (->>
       sources
-      (mapcat #(process-source-tree cl %))
+      (mapcat
+       #(process-source-tree (.getAbsolutePath output-path) cl %))
       (map #(mojo-descriptor plugin-descriptor %))))))