+++ plexus-clojure-compiler/src/it/compile/src/main/clojure/zitest/test.clj
@@ -1,0 +1,5 @@
+(ns zitest.test)
+
+(def x 1)
+
+(some-undefined)