+++ plexus-clojure-compiler/src/it/compile/src/main/clojure/zitest/test2.clj
@@ -1,0 +1,3 @@
+(ns zitest.test2)
+
+(some-undefined2)