+++ ReleaseNotes.md
@@ -1,6 +1,18 @@
 # Release Notes
 
-Current release is 0.1.0.
+Current release is 0.2.0.
+
+## 0.2.0
+
+- Compile clojure before looking for Mojos
+  Enables use of gen-class to build mojos.
+
+- Add plexus clojure compiler component
+
+- Add Phase annotation for mojos
+
+- Remove check on classloader of generated component
+
 
 ## 0.1.0